修复一定条件下dx截屏花屏问题（忽略rowPitch引起）

+++ op/include/Image.hpp
@@ -225,7 +225,7 @@ struct ImageBin {
 	void fromImage4(const Image& img4) {
 		create(img4.width, img4.height);
 		auto psrc = img4.pdata;
-		for (int i = 0; i < pixels.size(); ++i) {
+		for (size_t i = 0; i < pixels.size(); ++i) {
 			//pixels[i] = (psrc[0] + psrc[1] + psrc[2]) / 3;
 			// Gray = (R*299 + G*587 + B*114 + 500) / 1000
 			pixels[i] = (psrc[2] * 299 + psrc[1] * 587 + psrc[0] * 114 + 500) / 1000;
